requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.5
djangorestframework==3.14.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
gunicorn==21.2.0
h11==0.14.0
idna==3.4
numpy==1.23.0
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.0
psycopg2-binary==2.9.7
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
TgCrypto==1.2.5
trio==0.22.2
trio-websocket==0.10.4
tzdata==2023.3
urllib3==2.0.5
webdriver-manager==4.0.1
whitenoise==6.5.0
wsproto==1.2.0
